fix: play again button stutter after first replay
More specifically, fixed an issue where, after playing the game for a second time after using the Play Again button, when the menu would appear again, the Play Again button would appear in its active state, then it would jump to inactive and then transition to active again. The fix just added another method in the BaseUIButton to snap it into position instead of transitioning it.