fix: player input not working
The main cause of the bug was just method execution. Each prefab has the PlayerType enum set to CPU. When Object.Instantiate is called, Awake() and OnEnable() run. Since it is set to CPU by default, the condition on Character.cs:197 always returned true, meaning that the Attack performed event isn't being subscribed to. The temporary fix implemented here called that same condition *after* the instantiation