fix: standoff mode not starting after opening options menu
At the start of the game, opening the options menu, closing it, then trying to start standoff mode would cause a null ref exception since the first button to call DisableButton, which runs _button.enabled = false, is the Ronin button. It's 1:30 am, idk if I'm wording this right, but basically after opening and closing the options menu, the character buttons did not unsubscribe to the DisableButton event call. After the menu is closed, it is null, hence the null ref E. The solution was to unsubscribe from it in the OnDisable method of BaseUIButton instead of the OnDestroy. Another solution might have been adding a conditional to check if _button is null before setting its enabled to false, but this is better.