feat: shogun sets instantly without exit time
This implements the change where a character, if it is a Player, as soon as the third Haiku line is submitted, starts the Set animation right away without delay. This is meant to sync the Player's physical input closer to the animation. The CPU still has the original delay. This feature was implemented by adding an identical transition except for the exit time, and a new trigger for it. As of now, this only works with Shogun, but I'll implement it for the other Characters in near future commmits